64-bit The CRC headers now include gen_defs.h

+++ src/hash/crc16.c
@@ -38,7 +38,7 @@
 #include <string.h>	/* strlen */
 #include "crc16.h"
 
-unsigned short crc16tbl[] = {
+uint16_t crc16tbl[] = {
 0x0000, 0x1021, 0x2042, 0x3063, 0x4084, 0x50A5, 0x60C6, 0x70E7,
 0x8108, 0x9129, 0xA14A, 0xB16B, 0xC18C, 0xD1AD, 0xE1CE, 0xF1EF,
 0x1231, 0x0210, 0x3273, 0x2252, 0x52B5, 0x4294, 0x72F7, 0x62D6,
@@ -73,10 +73,10 @@ unsigned short crc16tbl[] = {
 0x6E17, 0x7E36, 0x4E55, 0x5E74, 0x2E93, 0x3EB2, 0x0ED1, 0x1EF0
 };
 
-unsigned short crc16(char* data, unsigned long len)
+uint16_t crc16(char* data, unsigned long len)
 {
-	unsigned short crc = 0;
-	unsigned long l;
+	uint16_t crc = 0;
+	uint32_t l;
 
 	if(len==0 && data!=NULL)
 		len=strlen(data);
