+++ src/smblib/smbadd.c
@@ -55,7 +55,7 @@ int SMBCALL smb_addmsg(smb_t* smb, smbmsg_t* msg, int storage, long dupechk_hash
 	size_t		bodylen=0;
 	size_t		chklen=0;
 	long		offset;
-	ulong		crc=0xffffffff;
+	uint32_t	crc=0xffffffff;
 	hash_t		found;
 	hash_t**	hashes=NULL;	/* This is a NULL-terminated list of hashes */
 	smbmsg_t	remsg;
